backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ian <exor2008@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"
numpy = "^1.26.3"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
